code_final/requirements.txt
jsonrpclib-pelix
newspaper3k
numpy
pandas
pika
pymongo
python-dateutil
requests
redis
scipy
sklearn
tensorflow
watchdog
